requirements.txt
pyrogram
tgcrypto

pymongo[srv]
motor
dnspython
aiosqlite

requests
aiohttp

environs

GitPython

wheel
pygments
ffmpeg-python
Pillow>=9.0.0
pytube
click
